Add AuthorizationError
