+++ tests/ApiApprovalTests/GraphQL.Server.Authorization.AspNetCore.approved.txt
@@ -1,5 +1,11 @@
 namespace GraphQL.Server.Authorization.AspNetCore
 {
+    public class AuthorizationError : GraphQL.Validation.ValidationError
+    {
+        public AuthorizationError(GraphQL.Language.AST.INode node, GraphQL.Validation.ValidationContext context, GraphQL.Language.AST.OperationType? operationType, Microsoft.AspNetCore.Authorization.AuthorizationResult result) { }
+        public AuthorizationError(GraphQL.Language.AST.INode node, GraphQL.Validation.ValidationContext context, string message, Microsoft.AspNetCore.Authorization.AuthorizationResult result) { }
+        public virtual Microsoft.AspNetCore.Authorization.AuthorizationResult AuthorizationResult { get; }
+    }
     public static class AuthorizationMetadataExtensions
     {
         public const string POLICY_KEY = "Authorization__Policies";
